Adjusting contracts after changing domain.. Adding dummies for it to build

+++ src/StockService/Domain/Entities/Location.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
     public class Location
     {
         public String Address { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }        
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }        
     }
 }
